[ADD] flexwheels: added demo data, access rights, constraints and graph view.
Flexwheels module was lacking demo data so defined a new demo data for it. Also
defined new access rights for this module and created two groups one for agent
and other for manager of system. Agent was given limited rights while manager
was provided with all rights. Added delete constraints to bookings to allow
deletion of only draft and cancelled bookings. Also added a new graph view in
booking view.