[ADD] flexwheels: added websited and improved handling of availability.
Improved handling of availability for booking by showing only those customers
and cars which are available in given period of booking. Added website to show
list of cars. Added kanban to customer view and improved kanban of booking and
car. Added Groupby and search in booking, car and customer. Updated demo data.